refactor: add base methods for processing and generating source text
- Introduced ProccessCandidates method to centralize request processing logic.
- Added GenerateSourceText for generating source text from requests.
- Refactored ExistingFieldPropertiesGenerator and NonExistingFieldPropertiesSyntaxGenerator to use new base methods.
