+++ PropertyBitPack.SourceGen/PropertiesSyntaxGenerators/ExistingFieldPropertiesGenerator.cs
@@ -24,48 +24,6 @@ internal sealed class ExistingFieldPropertiesGenerator: BasePropertiesSyntaxGene
     {
         var candidateRequests = FilterCandidates<ExistingFieldGsr>(requests);
 
-        for (var i = 0; i < candidateRequests.Length; i++)
-        {
-            var candidateRequest = candidateRequests[i];
-
-            var sourceText = Generate(candidateRequest);
-            var fileName = GetFileName(candidateRequest);
-
-            requests.Remove(candidateRequest);
-
-            immutableArrayBuilder.Add(new FileGeneratorRequest(sourceText, fileName));
-        }
-
-    }
-
-    private SourceText Generate(ExistingFieldGsr nonExistingFieldGsr)
-    {
-        using var propertiesRented = ListsPool.Rent<PropertyDeclarationSyntax>();
-        var properties = propertiesRented.List;
-
-        using var additionalMembersRented = ListsPool.Rent<MemberDeclarationSyntax>();
-        var additionalMembersList = additionalMembersRented.List;
-
-        for (var i = 0; i < nonExistingFieldGsr.Properties.Length; i++)
-        {
-            var propertyRequest = nonExistingFieldGsr.Properties[i];
-
-            properties.Add(GenerateProperty(nonExistingFieldGsr, propertyRequest, out var additionalMembers));
-
-            if (!additionalMembers.IsDefaultOrEmpty)
-            {
-                additionalMembersList.AddRange(additionalMembers);
-            }
-        }
-
-        using var membersRented = ListsPool.Rent<MemberDeclarationSyntax>();
-        var members = membersRented.List;
-
-        members.AddRange(properties);
-        members.AddRange(additionalMembersList);
-
-        var unit = GenerateCompilationUnit(nonExistingFieldGsr, members);
-
-        return unit.NormalizeWhitespace().GetText(Encoding.UTF8);
+        ProccessCandidates(requests, candidateRequests, immutableArrayBuilder);
     }
 }
