+++ PropertyBitPack.SourceGen/PropertiesSyntaxGenerators/NonExistingFieldPropertiesSyntaxGenerator.cs
@@ -22,61 +22,10 @@ namespace PropertyBitPack.SourceGen.PropertiesSyntaxGenerators;
 /// </remarks>
 internal sealed class NonExistingFieldPropertiesSyntaxGenerator : BasePropertiesSyntaxGenerator
 {
-    protected override void GenerateCore(ILinkedList<GenerateSourceRequest> requests, in ImmutableArrayBuilder<FileGeneratorRequest> immutableArrayBuilder)
+    protected override void GenerateCore(ILinkedList<GenerateSourceRequest> requests, in ImmutableArrayBuilder<FileGeneratorRequest> fileGeneratorRequestsBuilder)
     {
         var candidateRequests = FilterCandidates<NonExistingFieldGsr>(requests);
 
-        for (var i = 0; i < candidateRequests.Length; i++)
-        {
-            var candidateRequest = candidateRequests[i];
-
-            var sourceText = Generate(candidateRequest);
-            var fileName = GetFileName(candidateRequest);
-
-            requests.Remove(candidateRequest);
-
-            immutableArrayBuilder.Add(new FileGeneratorRequest(sourceText, fileName));
-        }
-
-    }
-
-    private SourceText Generate(NonExistingFieldGsr nonExistingFieldGsr)
-    {
-        using var fieldsRented = ListsPool.Rent<FieldDeclarationSyntax>();
-        var fields = fieldsRented.List;
-
-        foreach (var field in nonExistingFieldGsr.NonExistingFieldRequests)
-        {
-            fields.Add(GenerateField(nonExistingFieldGsr, field));
-        }
-
-        using var propertiesRented = ListsPool.Rent<PropertyDeclarationSyntax>();
-        var properties = propertiesRented.List;
-
-        using var additionalMembersRented = ListsPool.Rent<MemberDeclarationSyntax>();
-        var additionalMembersList = additionalMembersRented.List;
-
-        for (var i = 0; i < nonExistingFieldGsr.Properties.Length; i++)
-        {
-            var propertyRequest = nonExistingFieldGsr.Properties[i];
-
-            properties.Add(GenerateProperty(nonExistingFieldGsr, propertyRequest, out var additionalMembers));
-
-            if (!additionalMembers.IsDefaultOrEmpty)
-            {
-                additionalMembersList.AddRange(additionalMembers);
-            }
-        }
-
-        using var membersRented = ListsPool.Rent<MemberDeclarationSyntax>();
-        var members = membersRented.List;
-
-        members.AddRange(fields);
-        members.AddRange(properties);
-        members.AddRange(additionalMembersList);
-
-        var unit = GenerateCompilationUnit(nonExistingFieldGsr, members);
-
-        return unit.NormalizeWhitespace().GetText(Encoding.UTF8);
+        ProccessCandidates(requests, candidateRequests, fileGeneratorRequestsBuilder);
     }
 }
